Conflicts:
	ggEngine/ggEngine/PingPongState.cpp

+++ ggEngine/ggEngine/SpriteInfo.cpp
@@ -5,7 +5,7 @@ ggEngine::SpriteInfo::SpriteInfo(Texture * texture)
 	this->texture = texture;
 	this->width = texture->GetWidth();
 	this->height = texture->GetHeight();
-	this->srcRect = { 0,0,width, height };
+	this->srcRect = { 0,0,(long)width, (long)height };
 	this->type = SpriteType_Texture;
 }
 
@@ -14,7 +14,7 @@ ggEngine::SpriteInfo::SpriteInfo(Texture * atlas, float x, float y, float width,
 	this->texture = atlas;
 	this->width = width;
 	this->height = height;
-	this->srcRect = { x,y,x + width,y + height };
+	this->srcRect = { (long)x, (long)y, (long)(x + width),(long)(y + height) };
 	this->type = SpriteType_Atlas;
 }
